requirements.txt
fastapi[standard]
aiosqlite
pydantic
pydantic-ai-slim[groq]
